Enable ServiceAlertStatus
At the end of our reconciliation logic, we were attempting to update the
current status of our `ServiceAlerter` object. This was the cause of the
errors we were seeing, due to the status not actually being a part of
our CRD.

Defining the status resource for the kube macro fixes this bug.