Filter_map actually unnecessary (real this time)
The breakage I was seeing was a massive slow down when deployed
in-cluster, which actually was due to a permissions issue, where the
ClusterRole wasn't able to watch events or ConfigMaps.

The FilterMap in this case is just ocnverting Result<T, E> into
Option<T>, which was a no-op as far as I'm aware, and the kube.rs docs
never show this in their examples, so I'm happy to dump it.